Revert "Re-enable mac tests"
This reverts commit cc811f77fc4f48fb696784f0a51042bb736d7071.

GUI tests were temporarily disabled on Mac in 8ed38c7 and cc811f7 meant
to re-enable them, but in fact removed the GUI tests on every platform.
There is still plenty of instability in the tests, so we will first
restore GUI tests on non-Mac platforms, and then we'll try to properly
re-enable them on Mac, step by step.

Change-Id: Ic8ac32386cc9e7c5ea3172751d74138e30fa0fdf
Reviewed-by: Jan Arve Sæther <jan-arve.saether@digia.com>
Reviewed-by: Lars Knoll <lars.knoll@digia.com>