Fix origin for short reversed item views
Reversed (BottomToTop and RightToLeft) item views would report an
impossible positive value for origin when contentHeight < height. The
correct value will be effectively equivalent to the position of the
footer, and always negative.

Match other logic using lastPosition() by negating its value for
reversed views.

This fixes a bug with content disappearing in a Controls ScrollView on
a BottomToTop view.

Change-Id: Ieedbb64ce8fc7c0fb36e5256e437ddeb3e757761
Reviewed-by: Martin Jones <martin.jones@jollamobile.com>