Add some leading white space to generated code for QML signals.
Like this the column numbers in the generated code are correct. We rely
on the column numbers in quite a few places.

The alternative solution of extending the Javascript parser to add an
initial column offset would result in much more code and complexity.

const-casting the column number and adding to it after compiling would
be an option but as the column number may be saved in other places, too,
this is not desirable.

Task-number: QTBUG-37226
Change-Id: I0d11816d96ff4e7c756064646755a3e9874f9790
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>