PinchArea: zooming requires pinch distance to exceed the threshold
If you use a PinchArea and a Flickable together on OSX, you will be
able to use two fingers on the trackpad to flick the Flickable because
the OS will send scroll events which Qt turns into QWheelEvents; and
at the same time you will have two touch points, which the PinchArea
will grab and use.  That's nice, but it means we need to be more
selective about gesture recognition, so that when you mean to flick
you can do it without simultaneously pinch-zooming.
Therefore, iff pinch.dragAxis is Pinch.NoDrag, the startDragDistance
(threshold) is treated as a pinch distance threshold rather than
allowing any direction of movement to begin the pinch gesture.

[Changelog][QtQuick][PinchArea] do not begin zooming until the
change in distance between two touch points exceeds startDragDistance

Task-number: QTBUG-37294
Change-Id: I01026364970428aae9a4c5436517db719387f8e8
Reviewed-by: Laszlo Agocs <laszlo.agocs@digia.com>
Reviewed-by: Ulf Hermann <ulf.hermann@digia.com>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@digia.com>